pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asam-ods-exd-api-parquet"
version = "0.9.0"
description = "ASAM ODS EXD API implementation to read external data from parquet files."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{name = "totonga"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "grpcio>=1.76.0,<2.0.0",
    "grpcio-health-checking>=1.76.0,<2.0.0",
    "numpy>=2.4.1,<3.0.0",
    "pandas>=2.3.3,<3.0.0",
    "pyarrow>=22.0.0",
    "ods-exd-api-box>=1.0.1,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "coverage>=7.0",
    "isort>=5.12",
    "jupyterlab>=4.0",
    "mypy>=1.5",
    "types-protobuf",
    "types-grpcio",
    "pyarrow.stubs"
]

[project.urls]
Homepage = "https://github.com/totonga/asam_ods_exd_api_parquet"

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
]

[tool.pylint.format]
max-line-length = 109

[tool.black]
line-length = 109

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "ods_exd_api_box.proto.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ods_pb2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ods_external_data_pb2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ods_external_data_pb2_grpc"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "grpc"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyarrow.*"
ignore_missing_imports = true